Minor Fixes
An error was produce, if the LABLE is before the FROM then the first time the image is built an exception happens.
This change is fixing an issue when this project is run for the fist time need to have the container image build locally, adding formula under the package target so the image is always created.